Last edits made in March 2014
Note: This project was currently under development when the decision
was made to end my contract. The code herein was the product of
several, several hours of unpaid dedication to the amazing team I was
working for. Not all of the 3rd party technologies used were the best
solutions, but I did my best to keep the code looking clean, logical,
and to the point. I would love to pickup a project like this again in
the future.